separate code in functions and scripts
