+++ tests/test_apel_plugin.py
@@ -1,5 +1,5 @@
 import pytest
-import apel_plugin.apel_plugin as apel_plugin
+import apel_plugin.functions as functions
 
 
 @pytest.mark.asyncio
@@ -11,13 +11,13 @@ class TestAPELPlugin:
 
         dict = {"^a": "apple_in_dict", "^b": "banana_in_dict"}
 
-        result = await apel_plugin.regex_dict_lookup(term_a, dict)
+        result = await functions.regex_dict_lookup(term_a, dict)
         assert result == "apple_in_dict"
 
-        result = await apel_plugin.regex_dict_lookup(term_b, dict)
+        result = await functions.regex_dict_lookup(term_b, dict)
         assert result == "banana_in_dict"
 
         with pytest.raises(SystemExit) as pytest_error:
-            await apel_plugin.regex_dict_lookup(term_c, dict)
+            await functions.regex_dict_lookup(term_c, dict)
         assert pytest_error.type == SystemExit
         assert pytest_error.value.code == 1
